requirements.txt
flask
twilio
pyautogui
pyparsing
pyperclip
requests
langdetect
deep-translator
pyaudio
google-generativeai
